Added updating functions to all parts of the game
Those should be used to update the proper objects on the client side. It remains
to create the encoding function for the server.

This is actually a bit of a hack and everything needs a bit of a rework. The
client-side objects should probably be of a different class than the server
side, because they should be doing different things (maybe more M and C in
MVC?). Another problem that has surfaced is reffering to tiles by pointers --
maybe we should have just one global vector with all the tiles and refer to them
by some id? All that after Eduardo, unless The Others say otherwise.