Server can now be started
Added a network contestant for that to be possible. This resulted in some
changes all around. Most interesting ones:
* encoding as DFStyle is now const
* some changes in requesting decisions from the contestants
* more stuff can be encoded as DFStyle

This also means that after everybody finishes what they were supposed to do we
will have a working prototype.

Some problematic code has been introduced, especially as far as the
communication protocol goes, but this will probably be fixed after Edziocalypse.